requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.2
confection==0.1.4
cymem==2.0.8
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.13.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
langcodes==3.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
multidict==6.0.5
murmurhash==1.0.10
nest-asyncio==1.6.0
numpy==1.26.4
openai==0.28.0
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
parso==0.8.3
platformdirs==4.2.0
preshed==3.0.9
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
thinc==8.2.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typer==0.9.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
yarl==1.9.4
